sql files were added

+++ src/main/resources/db/table/department_type.sql
@@ -3,6 +3,6 @@
 --changeset anon:department_type-1
 
 create table online_manager.department_type(
-        id bigserial not null primary key,
+        id int not null primary key,
         name varchar(50) not null unique
 );
